+++ contracts/Migrations.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 
 contract Migrations {
     address public _owner;
-    uint public _last_completed_migration;
+    uint256 public _last_completed_migration;
 
     constructor() {
         _owner = msg.sender;
@@ -14,7 +14,7 @@ contract Migrations {
         if (msg.sender == _owner) _;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         _last_completed_migration = completed;
     }
